whitelabel-rag/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.5.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
werkzeug>=2.3.0
gunicorn>=21.0.0
redis>=5.0.0
pytest>=7.4.0
pytest-flask>=1.2.0
factory-boy>=3.3.0